Split main implementation file into multiple files and tests
