+++ src/lib/workflow/definition.test.ts
@@ -1,652 +1,17 @@
 import { beforeEach, describe, expect, test } from "bun:test";
+import { clearWorkflowRegistry } from "./index.js";
 import {
-  clearWorkflowRegistry,
-  defineWorkflow as defineWorkflowActual,
-  listWorkflows,
-  recoverOrphanedWorkflows,
-  resumeWorkflow,
-} from "./index.js";
-import type { WorkflowOptions } from "./types.js";
-
-const defineWorkflow = <TInput, TResult = void>(
-  options: WorkflowOptions<TInput, TResult>,
-) => {
-  return defineWorkflowActual<TInput, TResult>({
-    recoveryIntervalMs: 0,
-    ...options,
-  });
-};
-
-class MockRedisClient {
-  private hashes = new Map<string, Map<string, string>>();
-  private strings = new Map<string, string>();
-  private lists = new Map<string, string[]>();
-  private expirations = new Map<string, number>();
-  private failCommands = new Set<string>();
-  private hsetCallsBeforeFail: number | null = null;
-  public hsetFailureCount = 0;
-
-  private isExpired(key: string) {
-    const expiry = this.expirations.get(key);
-
-    if (expiry === undefined) {
-      return false;
-    }
-
-    if (Date.now() >= expiry) {
-      this.strings.delete(key);
-      this.expirations.delete(key);
-      return true;
-    }
-
-    return false;
-  }
-
-  private matchesGlob(key: string, pattern: string) {
-    const escaped = pattern.replace(/[.+^${}()|[\]\\]/g, "\\$&");
-    const regex = new RegExp(`^${escaped.replace(/\*/g, ".*")}$`);
-
-    return regex.test(key);
-  }
-
-  public setCommandFailure(
-    command:
-      | "hset"
-      | "hget"
-      | "set"
-      | "get"
-      | "del"
-      | "lpush"
-      | "rpop"
-      | "rpoplpush"
-      | "lrem"
-      | "lrange"
-      | "eval",
-  ) {
-    this.failCommands.add(command);
-  }
-
-  public failHsetAfterNCalls(n: number) {
-    this.hsetCallsBeforeFail = n;
-  }
-
-  public seedHashField(key: string, field: string, value: string) {
-    if (!this.hashes.has(key)) {
-      this.hashes.set(key, new Map());
-    }
-
-    this.hashes.get(key)?.set(field, value);
-  }
-
-  public async hset(key: string, field: string, value: string): Promise<number>;
-
-  public async hset(
-    key: string,
-    values: Record<string, string>,
-  ): Promise<number>;
-
-  public async hset(
-    key: string,
-    fieldOrValues: string | Record<string, string>,
-    value?: string,
-  ) {
-    if (this.failCommands.has("hset")) {
-      this.hsetFailureCount++;
-      throw new Error("hset failed");
-    }
-
-    if (this.hsetCallsBeforeFail !== null) {
-      if (this.hsetCallsBeforeFail <= 0) {
-        this.hsetFailureCount++;
-        throw new Error("hset failed");
-      }
-
-      this.hsetCallsBeforeFail--;
-    }
-
-    if (!this.hashes.has(key)) {
-      this.hashes.set(key, new Map());
-    }
-
-    const hash = this.hashes.get(key);
-
-    if (!hash) {
-      return 0;
-    }
-
-    if (typeof fieldOrValues === "string") {
-      if (typeof value !== "string") {
-        return 0;
-      }
-
-      hash.set(fieldOrValues, value);
-      return 1;
-    }
-
-    let count = 0;
-
-    for (const [field, entry] of Object.entries(fieldOrValues)) {
-      if (!hash.has(field)) {
-        count++;
-      }
-
-      hash.set(field, entry);
-    }
-
-    return count;
-  }
-
-  public async hget(key: string, field: string) {
-    if (this.failCommands.has("hget")) {
-      throw new Error("hget failed");
-    }
-
-    const hash = this.hashes.get(key);
-
-    if (!hash) {
-      return null;
-    }
-
-    return hash.get(field) ?? null;
-  }
-
-  public async hgetall(key: string) {
-    const hash = this.hashes.get(key);
-
-    if (!hash) {
-      return {};
-    }
-
-    return Object.fromEntries(hash.entries());
-  }
-
-  public async exists(key: string) {
-    if (this.isExpired(key)) {
-      return false;
-    }
-
-    if (this.strings.has(key)) {
-      return true;
-    }
-
-    if (this.hashes.has(key)) {
-      return true;
-    }
-
-    return false;
-  }
-
-  public async scan(
-    cursor: string | number,
-    ...args: (string | number)[]
-  ): Promise<[string, string[]]> {
-    let pattern: string | null = null;
-
-    for (let i = 0; i < args.length; i++) {
-      if (args[i] === "MATCH" && typeof args[i + 1] === "string") {
-        pattern = args[i + 1] as string;
-      }
-    }
-
-    const keys: string[] = [];
-
-    for (const key of this.hashes.keys()) {
-      if (pattern) {
-        if (!this.matchesGlob(key, pattern)) {
-          continue;
-        }
-      }
-
-      keys.push(key);
-    }
-
-    for (const key of this.strings.keys()) {
-      if (this.isExpired(key)) {
-        continue;
-      }
-
-      if (pattern) {
-        if (!this.matchesGlob(key, pattern)) {
-          continue;
-        }
-      }
-
-      keys.push(key);
-    }
-
-    // ponytail: mock returns all matches in one page
-    if (String(cursor) !== "0") {
-      return ["0", []];
-    }
-
-    return ["0", keys];
-  }
-
-  public async set(key: string, value: string, ...args: unknown[]) {
-    if (this.failCommands.has("set")) {
-      throw new Error("set failed");
-    }
-
-    let nx = false;
-    let xx = false;
-    let pxValue: number | null = null;
-
-    for (let i = 0; i < args.length; i++) {
-      if (args[i] === "NX") {
-        nx = true;
-      }
-
-      if (args[i] === "XX") {
-        xx = true;
-      }
-
-      if (args[i] === "PX" && typeof args[i + 1] === "string") {
-        pxValue = parseInt(args[i + 1] as string, 10);
-      }
-    }
-
-    this.isExpired(key);
-
-    const exists = this.strings.has(key);
-
-    if (nx && exists) {
-      return null;
-    }
-
-    if (xx && !exists) {
-      return null;
-    }
-
-    this.strings.set(key, value);
-
-    if (pxValue !== null) {
-      this.expirations.set(key, Date.now() + pxValue);
-    } else {
-      this.expirations.delete(key);
-    }
-
-    return "OK";
-  }
-
-  public async get(key: string) {
-    if (this.failCommands.has("get")) {
-      throw new Error("get failed");
-    }
-
-    if (this.isExpired(key)) {
-      return null;
-    }
-
-    return this.strings.get(key) ?? null;
-  }
-
-  public async del(key: string) {
-    if (this.failCommands.has("del")) {
-      throw new Error("del failed");
-    }
-
-    let count = 0;
-
-    if (this.strings.delete(key)) {
-      count++;
-    }
-
-    if (this.hashes.delete(key)) {
-      count++;
-    }
-
-    if (this.lists.delete(key)) {
-      count++;
-    }
-
-    return count;
-  }
-
-  public async lpush(key: string, value: string) {
-    if (this.failCommands.has("lpush")) {
-      throw new Error("lpush failed");
-    }
-
-    if (!this.lists.has(key)) {
-      this.lists.set(key, []);
-    }
-
-    this.lists.get(key)?.unshift(value);
-
-    return this.lists.get(key)?.length ?? 0;
-  }
-
-  public async rpop(key: string) {
-    if (this.failCommands.has("rpop")) {
-      throw new Error("rpop failed");
-    }
-
-    const list = this.lists.get(key);
-
-    if (!list || list.length === 0) {
-      return null;
-    }
-
-    return list.pop() ?? null;
-  }
-
-  public async rpoplpush(source: string, destination: string) {
-    if (this.failCommands.has("rpoplpush")) {
-      throw new Error("rpoplpush failed");
-    }
-
-    const value = await this.rpop(source);
-
-    if (value === null) {
-      return null;
-    }
-
-    if (!this.lists.has(destination)) {
-      this.lists.set(destination, []);
-    }
-
-    this.lists.get(destination)?.unshift(value);
-
-    return value;
-  }
-
-  public async lrem(key: string, count: number, value: string) {
-    if (this.failCommands.has("lrem")) {
-      throw new Error("lrem failed");
-    }
-
-    const list = this.lists.get(key);
-
-    if (!list) {
-      return 0;
-    }
-
-    let removed = 0;
-
-    if (count === 0) {
-      const next = list.filter((entry) => {
-        if (entry === value) {
-          removed++;
-          return false;
-        }
-
-        return true;
-      });
-
-      this.lists.set(key, next);
-      return removed;
-    }
-
-    if (count > 0) {
-      const next: string[] = [];
-
-      for (const entry of list) {
-        if (entry === value && removed < count) {
-          removed++;
-          continue;
-        }
-
-        next.push(entry);
-      }
-
-      this.lists.set(key, next);
-      return removed;
-    }
-
-    const next: string[] = [];
-    const limit = Math.abs(count);
-
-    for (let i = list.length - 1; i >= 0; i--) {
-      const entry = list[i];
-
-      if (entry === value && removed < limit) {
-        removed++;
-        continue;
-      }
-
-      next.unshift(entry ?? "");
-    }
-
-    this.lists.set(key, next);
-    return removed;
-  }
-
-  public async lrange(key: string, start: number, stop: number) {
-    if (this.failCommands.has("lrange")) {
-      throw new Error("lrange failed");
-    }
-
-    const list = this.lists.get(key) ?? [];
-
-    if (list.length === 0) {
-      return [] as string[];
-    }
-
-    let from = start;
-    let to = stop;
-
-    if (from < 0) {
-      from = list.length + from;
-    }
-
-    if (to < 0) {
-      to = list.length + to;
-    }
-
-    if (from < 0) {
-      from = 0;
-    }
-
-    if (to >= list.length) {
-      to = list.length - 1;
-    }
-
-    if (from > to) {
-      return [] as string[];
-    }
-
-    return list.slice(from, to + 1);
-  }
-
-  public seedList(key: string, values: string[]) {
-    this.lists.set(key, [...values]);
-  }
-
-  public getList(key: string) {
-    return [...(this.lists.get(key) ?? [])];
-  }
-
-  public async send(command: string, args: string[]) {
-    if (command !== "EVAL") {
-      throw new Error(`Unsupported command: ${command}`);
-    }
-
-    if (this.failCommands.has("eval")) {
-      throw new Error("eval failed");
-    }
-
-    const script = args[0];
-    const numKeys = parseInt(args[1] ?? "0", 10);
-    const keys = args.slice(2, 2 + numKeys);
-    const argv = args.slice(2 + numKeys);
-
-    if (!script) {
-      throw new Error("EVAL requires a script");
-    }
-
-    // releaseLock: GET compare then DEL
-    if (script.includes("'DEL'")) {
-      if (this.isExpired(keys[0] ?? "")) {
-        return 0;
-      }
-
-      const current = this.strings.get(keys[0] ?? "");
-
-      if (current === argv[0]) {
-        this.strings.delete(keys[0] ?? "");
-        this.expirations.delete(keys[0] ?? "");
-        return 1;
-      }
-
-      return 0;
-    }
-
-    // extendLock: GET compare then PEXPIRE
-    if (script.includes("'PEXPIRE'")) {
-      if (this.isExpired(keys[0] ?? "")) {
-        return 0;
-      }
-
-      const current = this.strings.get(keys[0] ?? "");
-
-      if (current === argv[0]) {
-        const ms = parseInt(argv[1] ?? "0", 10);
-        this.expirations.set(keys[0] ?? "", Date.now() + ms);
-        return 1;
-      }
-
-      return 0;
-    }
-
-    // createExecution: EXISTS check then HSET all fields
-    if (script.includes("'EXISTS'") && script.includes("'HSET'")) {
-      const metaKey = keys[0] ?? "";
-
-      if (this.hashes.has(metaKey)) {
-        return 0;
-      }
-
-      const hash = new Map<string, string>();
-
-      for (let i = 0; i + 1 < argv.length; i += 2) {
-        hash.set(argv[i] ?? "", argv[i + 1] ?? "");
-      }
-
-      this.hashes.set(metaKey, hash);
-      return 1;
-    }
-
-    throw new Error("Unknown EVAL script");
-  }
-}
-
-const createRedis = () => {
-  return new MockRedisClient() as MockRedisClient & Bun.RedisClient;
-};
-
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
-
-const waitForExecution = async <TInput, TResult>(
-  workflow: ReturnType<typeof defineWorkflow<TInput, TResult>>,
-  executionId: string,
-) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    const execution = await workflow.get(executionId);
-
-    if (execution.status !== "pending" && execution.status !== "running") {
-      return execution;
-    }
-
-    await sleep(1);
-  }
-
-  throw new Error(`Workflow execution ${executionId} did not finish`);
-};
-
-const waitForHsetFailure = async (redis: MockRedisClient, failureCount = 1) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    if (redis.hsetFailureCount >= failureCount) {
-      return;
-    }
-
-    await sleep(1);
-  }
-
-  throw new Error("Mock Redis hset did not fail");
-};
-
-const waitForStatus = async <TInput, TResult>(
-  workflow: ReturnType<typeof defineWorkflow<TInput, TResult>>,
-  executionId: string,
-  status: "running" | "completed" | "failed" | "cancelled",
-) => {
-  for (let attempt = 0; attempt < 3000; attempt++) {
-    const execution = await workflow.get(executionId);
-
-    if (execution.status === status) {
-      return execution;
-    }
-
-    await sleep(1);
-  }
-
-  throw new Error(`Workflow execution ${executionId} did not become ${status}`);
-};
-
-const fastRetryBackoff = {
-  baseDelay: 1,
-  multiplier: 2,
-  maxDelay: 10,
-};
-
-const createWorkflowWithEchoResult = (
-  name: string,
-  redis: Bun.RedisClient,
-  callCounter?: { value: number },
-) => {
-  return defineWorkflow<{ id: number }, string>({
-    name,
-    redis,
-    pollInterval: 1,
-    recoveryIntervalMs: 0,
-    handler: async ({ input, step }) => {
-      if (callCounter) {
-        callCounter.value++;
-      }
-
-      const value = await step("echo", async () => {
-        return `echo:${input.id}`;
-      });
-
-      return value;
-    },
-  });
-};
-
-const createTwoStepFailResumeWorkflow = (
-  name: string,
-  redis: Bun.RedisClient,
-  executedSteps: string[],
-) => {
-  let shouldFail = true;
-
-  return defineWorkflow<{ id: number }, string>({
-    name,
-    redis,
-    retries: 0,
-    pollInterval: 1,
-    recoveryIntervalMs: 0,
-    handler: async ({ input, step }) => {
-      await step("step-1", async () => {
-        executedSteps.push(`step-1:${input.id}`);
-        return "ok";
-      });
-
-      await step("step-2", async () => {
-        executedSteps.push(`step-2:${input.id}`);
-
-        if (shouldFail) {
-          shouldFail = false;
-          throw new Error("crash");
-        }
-
-        return "ok";
-      });
-
-      return "done";
-    },
-  });
-};
+  createRedis,
+  createTwoStepFailResumeWorkflow,
+  createWorkflowWithEchoResult,
+  defineWorkflow,
+  fastRetryBackoff,
+  type MockRedisClient,
+  sleep,
+  waitForExecution,
+  waitForHsetFailure,
+  waitForStatus,
+} from "./test-helpers.js";
 
 describe("workflow", () => {
   beforeEach(async () => {
@@ -1018,69 +383,45 @@ describe("workflow", () => {
     expect(typeof execution.steps[1]?.completedAt).toBe("number");
   });
 
-  describe("start matrix", () => {
-    for (let i = 1; i <= 20; i++) {
-      test(`starts workflow with id ${i}`, async () => {
-        const redis = createRedis();
-        const workflow = createWorkflowWithEchoResult(`run-matrix-${i}`, redis);
+  test("starts workflow with numeric id", async () => {
+    const redis = createRedis();
+    const workflow = createWorkflowWithEchoResult("run-matrix", redis);
 
-        await workflow.start(
-          { id: i },
-          { executionId: `run-matrix-exec-${i}` },
-        );
-        const execution = await waitForExecution(
-          workflow,
-          `run-matrix-exec-${i}`,
-        );
+    await workflow.start({ id: 7 }, { executionId: "run-matrix-exec" });
+    const execution = await waitForExecution(workflow, "run-matrix-exec");
 
-        expect(execution.result).toBe(`echo:${i}`);
-      });
-    }
+    expect(execution.result).toBe("echo:7");
   });
 
-  describe("resume matrix", () => {
-    for (let i = 1; i <= 15; i++) {
-      test(`does not re-run completed execution ${i}`, async () => {
-        const redis = createRedis();
-        const calls = { value: 0 };
-        const workflow = createWorkflowWithEchoResult(
-          `resume-matrix-${i}`,
-          redis,
-          calls,
-        );
+  test("does not re-run completed execution on resume", async () => {
+    const redis = createRedis();
+    const calls = { value: 0 };
+    const workflow = createWorkflowWithEchoResult(
+      "resume-matrix",
+      redis,
+      calls,
+    );
 
-        const executionId = `resume-matrix-exec-${i}`;
+    await workflow.start({ id: 3 }, { executionId: "resume-matrix-exec" });
+    await waitForStatus(workflow, "resume-matrix-exec", "completed");
 
-        await workflow.start({ id: i }, { executionId });
-        await waitForStatus(workflow, executionId, "completed");
+    const resumeData = await workflow.resume("resume-matrix-exec");
+    const execution = await workflow.get("resume-matrix-exec");
 
-        const resumeData = await workflow.resume(executionId);
-        const execution = await workflow.get(executionId);
-
-        expect(resumeData.status).toBe("completed");
-        expect(execution.result).toBe(`echo:${i}`);
-        expect(calls.value).toBe(1);
-      });
-    }
+    expect(resumeData.status).toBe("completed");
+    expect(execution.result).toBe("echo:3");
+    expect(calls.value).toBe(1);
   });
 
-  describe("duplicate matrix", () => {
-    for (let i = 1; i <= 10; i++) {
-      test(`rejects duplicate execution id case ${i}`, async () => {
-        const redis = createRedis();
-        const workflow = createWorkflowWithEchoResult(
-          `dupe-matrix-${i}`,
-          redis,
-        );
-        const executionId = `dupe-matrix-exec-${i}`;
+  test("rejects duplicate execution id", async () => {
+    const redis = createRedis();
+    const workflow = createWorkflowWithEchoResult("dupe-matrix", redis);
 
-        await workflow.start({ id: i }, { executionId });
+    await workflow.start({ id: 1 }, { executionId: "dupe-matrix-exec" });
 
-        await expect(
-          workflow.start({ id: i + 100 }, { executionId }),
-        ).rejects.toMatchObject({ name: "StateError" });
-      });
-    }
+    await expect(
+      workflow.start({ id: 2 }, { executionId: "dupe-matrix-exec" }),
+    ).rejects.toMatchObject({ name: "StateError" });
   });
 
   describe("invalid status matrix", () => {
@@ -2081,248 +1422,6 @@ describe("workflow", () => {
     });
   });
 
-  describe("listWorkflows and resumeWorkflow", () => {
-    test("returns empty list when no executions exist", async () => {
-      const redis = createRedis();
-
-      defineWorkflow<{ id: number }, string>({
-        name: "list-empty",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const listed = await listWorkflows(redis);
-
-      expect(listed).toEqual([]);
-    });
-
-    test("lists executions across multiple workflow names", async () => {
-      const redis = createRedis();
-
-      const onboard = defineWorkflow<{ id: number }, string>({
-        name: "list-onboard",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const provision = defineWorkflow<{ id: number }, string>({
-        name: "list-provision",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await onboard.start({ id: 1 }, { executionId: "list-onboard-1" });
-      await provision.start({ id: 2 }, { executionId: "list-provision-1" });
-      await waitForExecution(onboard, "list-onboard-1");
-      await waitForExecution(provision, "list-provision-1");
-
-      const listed = await listWorkflows(redis);
-      const names = listed.map((item) => item.name).sort();
-      const ids = listed.map((item) => item.id).sort();
-
-      expect(listed.length).toBe(2);
-      expect(names).toEqual(["list-onboard", "list-provision"]);
-      expect(ids).toEqual(["list-onboard-1", "list-provision-1"]);
-    });
-
-    test("filters by name and status", async () => {
-      const redis = createRedis();
-
-      const onboard = defineWorkflow<{ id: number }, string>({
-        name: "filter-onboard",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const provision = defineWorkflow<{ id: number }, string>({
-        name: "filter-provision",
-        redis,
-        retries: 0,
-        handler: async ({ step }) => {
-          await step("fail", async () => {
-            throw new Error("boom");
-          });
-
-          return "ok";
-        },
-      });
-
-      await onboard.start({ id: 1 }, { executionId: "filter-onboard-1" });
-      await provision.start({ id: 2 }, { executionId: "filter-provision-1" });
-      await waitForExecution(onboard, "filter-onboard-1");
-      await waitForStatus(provision, "filter-provision-1", "failed");
-
-      const byName = await listWorkflows(redis, { name: "filter-onboard" });
-      const byStatus = await listWorkflows(redis, { status: "failed" });
-
-      expect(byName.map((item) => item.id)).toEqual(["filter-onboard-1"]);
-      expect(byStatus.map((item) => item.id)).toEqual(["filter-provision-1"]);
-    });
-
-    test("unlockedOnly excludes locked running executions", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "list-lock",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "list-lock-held" });
-      await workflow.start({ id: 2 }, { executionId: "list-lock-orphan" });
-      await waitForExecution(workflow, "list-lock-held");
-      await waitForExecution(workflow, "list-lock-orphan");
-
-      redis.seedHashField(
-        "workflow:execution:list-lock-held:meta",
-        "status",
-        "running",
-      );
-      redis.seedHashField(
-        "workflow:execution:list-lock-orphan:meta",
-        "status",
-        "running",
-      );
-      await redis.set(
-        "workflow:execution:list-lock-held:lock",
-        "token",
-        "PX",
-        "60000",
-      );
-
-      const unlocked = await listWorkflows(redis, {
-        status: "running",
-        unlockedOnly: true,
-      });
-
-      expect(unlocked.map((item) => item.id)).toEqual(["list-lock-orphan"]);
-    });
-
-    test("resumeWorkflow resumes by execution id via registry", async () => {
-      const redis = createRedis();
-      const executedSteps: string[] = [];
-
-      const workflow = createTwoStepFailResumeWorkflow(
-        "resume-workflow-api",
-        redis,
-        executedSteps,
-      );
-
-      await workflow.start({ id: 10 }, { executionId: "resume-api-1" });
-      await waitForStatus(workflow, "resume-api-1", "failed");
-
-      const pending = await listWorkflows(redis, {
-        status: ["pending", "running", "failed"],
-        unlockedOnly: true,
-      });
-
-      expect(pending.some((item) => item.id === "resume-api-1")).toBe(true);
-
-      const resumed = await resumeWorkflow(redis, "resume-api-1");
-      const execution = await waitForStatus(
-        workflow,
-        "resume-api-1",
-        "completed",
-      );
-
-      expect(resumed.status).toBe("pending");
-      expect(execution.status).toBe("completed");
-      expect(executedSteps).toEqual(["step-1:10", "step-2:10", "step-2:10"]);
-    });
-
-    test("resumeWorkflow fails when workflow is not registered", async () => {
-      const redis = createRedis();
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "resume-unregistered",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "resume-unreg-1" });
-      await waitForExecution(workflow, "resume-unreg-1");
-
-      await clearWorkflowRegistry();
-
-      await expect(
-        resumeWorkflow(redis, "resume-unreg-1"),
-      ).rejects.toMatchObject({
-        name: "NotFoundError",
-        message:
-          "Workflow resume-unregistered is not registered in this process",
-      });
-    });
-
-    test("rejects duplicate workflow name registration", () => {
-      const redis = createRedis();
-
-      defineWorkflow<{ id: number }, string>({
-        name: "dupe-register",
-        redis,
-        handler: async () => "ok",
-      });
-
-      expect(() => {
-        defineWorkflow<{ id: number }, string>({
-          name: "dupe-register",
-          redis,
-          handler: async () => "ok",
-        });
-      }).toThrow(
-        "Workflow dupe-register is already registered in this process",
-      );
-    });
-
-    test("rejects custom execution id collision across workflow names", async () => {
-      const redis = createRedis();
-
-      const first = defineWorkflow<{ id: number }, string>({
-        name: "collision-a",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const second = defineWorkflow<{ id: number }, string>({
-        name: "collision-b",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await first.start({ id: 1 }, { executionId: "shared-id" });
-
-      await expect(
-        second.start({ id: 2 }, { executionId: "shared-id" }),
-      ).rejects.toMatchObject({
-        name: "StateError",
-        message: "Workflow execution shared-id already exists",
-      });
-    });
-
-    test("get rejects execution owned by another workflow name", async () => {
-      const redis = createRedis();
-
-      const first = defineWorkflow<{ id: number }, string>({
-        name: "owner-a",
-        redis,
-        handler: async () => "ok",
-      });
-
-      const second = defineWorkflow<{ id: number }, string>({
-        name: "owner-b",
-        redis,
-        handler: async () => "ok",
-      });
-
-      await first.start({ id: 1 }, { executionId: "owned-by-a" });
-      await waitForExecution(first, "owned-by-a");
-
-      await expect(second.get("owned-by-a")).rejects.toMatchObject({
-        name: "NotFoundError",
-        message: "Workflow execution owned-by-a not found",
-      });
-    });
-  });
-
   test("respects concurrency limits", async () => {
     const redis = createRedis();
     let maxConcurrent = 0;
@@ -2658,103 +1757,6 @@ describe("workflow", () => {
 
       expect(redis.getList("workflow:claim-ack:jobs")).toEqual([]);
       expect(redis.getList("workflow:claim-ack:processing")).toEqual([]);
-    });
-
-    test("recoverOrphanedWorkflows requeues processing orphans", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "recover-processing",
-        redis,
-        pollInterval: 1,
-        handler: async ({ step }) => {
-          await step("work", async () => "ok");
-          return "done";
-        },
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "recover-proc-1" });
-      await waitForExecution(workflow, "recover-proc-1");
-
-      redis.seedHashField(
-        "workflow:execution:recover-proc-1:meta",
-        "status",
-        "pending",
-      );
-      redis.seedHashField(
-        "workflow:execution:recover-proc-1:meta",
-        "completedAt",
-        "",
-      );
-      redis.seedHashField(
-        "workflow:execution:recover-proc-1:meta",
-        "result",
-        "",
-      );
-      redis.seedList("workflow:recover-processing:processing", [
-        "recover-proc-1",
-      ]);
-
-      const recovered = await recoverOrphanedWorkflows(redis, {
-        name: "recover-processing",
-      });
-
-      expect(recovered).toContain("recover-proc-1");
-
-      const execution = await waitForStatus(
-        workflow,
-        "recover-proc-1",
-        "completed",
-      );
-
-      expect(execution.status).toBe("completed");
-      expect(redis.getList("workflow:recover-processing:processing")).toEqual(
-        [],
-      );
-    });
-
-    test("recoverOrphanedWorkflows resumes unlocked pending without processing entry", async () => {
-      const redis = createRedis() as MockRedisClient & Bun.RedisClient;
-
-      const workflow = defineWorkflow<{ id: number }, string>({
-        name: "recover-orphan",
-        redis,
-        pollInterval: 1,
-        handler: async () => "done",
-      });
-
-      await workflow.start({ id: 1 }, { executionId: "recover-orphan-1" });
-      await waitForExecution(workflow, "recover-orphan-1");
-
-      redis.seedHashField(
-        "workflow:execution:recover-orphan-1:meta",
-        "status",
-        "pending",
-      );
-      redis.seedHashField(
-        "workflow:execution:recover-orphan-1:meta",
-        "completedAt",
-        "",
-      );
-      redis.seedHashField(
-        "workflow:execution:recover-orphan-1:meta",
-        "result",
-        "",
-      );
-
-      const recovered = await recoverOrphanedWorkflows(redis, {
-        name: "recover-orphan",
-      });
-
-      expect(recovered).toContain("recover-orphan-1");
-
-      const execution = await waitForStatus(
-        workflow,
-        "recover-orphan-1",
-        "completed",
-      );
-
-      expect(execution.status).toBe("completed");
     });
 
     test("partitionConcurrency caps per key independently of worker concurrency", async () => {
